DOC: Do not run nbgenerate if paths need updating.

+++ tools/nbgenerate.py
@@ -26,13 +26,18 @@ except ImportError:
 from IPython.nbformat.current import reads, write, NotebookNode
 
 cur_dir = os.path.abspath(os.path.dirname(__file__))
-
+gi
 # for conversion of .ipynb -> html/rst
 
 from IPython.config import Config
-if not os.path.exists(os.path.join(cur_dir, "nbconvert")):
-    os.symlink("/home/skipper/src/nbconvert",
+if os.path.exists("/home/skipper/src/nbconvert"):
+    if not os.path.exists(os.path.join(cur_dir, "nbconvert")):
+        os.symlink("/home/skipper/src/nbconvert",
                os.path.join(cur_dir, "nbconvert"))
+else:
+    from warnings import warn
+    warn("Notebook examples not built. Update paths in tools/nbgenerate.py")
+    sys.exit(0)
 from nbconvert.converters.template import ConverterTemplate
 from nbconvert.converters.rst import ConverterRST
